+++ src/components/events/EventPage.vue
