+++ src/components/projects/ProjectInfoBanner.vue
