+++ src/components/projects/index.ts
@@ -1,4 +1,4 @@
 export { default as ProjectsPage } from './ProjectsPage.vue'
-export { default as ProjectsInfoBanner } from './ProjectsInfoBanner.vue'
-export { default as ProjectCard } from './ProjectsCard.vue'
+export { default as ProjectsInfoBanner } from './ProjectInfoBanner.vue'
+export { default as ProjectCard } from './ProjectCard.vue'
 export { default as InterestModal } from './InterestModal.vue'
