+++ src/main.ts
@@ -91,8 +91,8 @@ import TeachingPage from '@/components/teaching/TeachingPage.vue'
 import CourseCard from '@/components/teaching/CourseCard.vue'
 
 import ProjectsPage from '@/components/projects/ProjectsPage.vue'
-import ProjectsInfoBanner from '@/components/projects/ProjectsInfoBanner.vue'
-import ProjectCard from '@/components/projects/ProjectsCard.vue'
+import ProjectsInfoBanner from '@/components/projects/ProjectInfoBanner.vue'
+import ProjectCard from '@/components/projects/ProjectCard.vue'
 import InterestModal from '@/components/projects/InterestModal.vue'
 
 import VacanciesPage from '@/components/vacancies/VacanciesPage.vue'
